renamed pydra.mark.shell_commands to pydra.mark.shell
