+++ pydra/mark/__init__.py
@@ -1,4 +1,4 @@
 from .functions import annotate, task
-from .shell_commands import shell_task, shell_arg, shell_out
+from .shell import shell_task, shell_arg, shell_out
 
 __all__ = ("annotate", "task", "shell_task", "shell_arg", "shell_out")
